refactor(number-input): ♻️  spin button hook
