+++ src/number-input/NumberInputDecrementButton.ts
@@ -1,8 +1,7 @@
-import { useCallback, useState } from "react";
 import { createComponent, createHook } from "reakit-system";
-import { ariaAttr, callAllHandlers } from "@chakra-ui/utils";
 import { ButtonHTMLProps, ButtonOptions, useButton } from "reakit/Button";
 
+import { useSpinButton } from "./__utils";
 import { NumberInputStateReturn } from "./NumberInputState";
 import { NUMBERINPUT_DECREMENTBUTTON_KEYS } from "./__keys";
 
@@ -27,56 +26,7 @@ export const useNumberInputDecrementButton = createHook<
   keys: NUMBERINPUT_DECREMENTBUTTON_KEYS,
 
   useProps(options, htmlProps) {
-    const {
-      keepWithinRange,
-      focusInput,
-      isAtMin: isAtMinProp,
-      spinner,
-    } = options;
-    const {
-      onMouseDown,
-      onMouseUp,
-      onMouseLeave,
-      onTouchStart,
-      onTouchEnd,
-      ...restHtmlProps
-    } = htmlProps;
-
-    const [isAtMin, setIsAtMin] = useState(false);
-
-    const spinDown = useCallback(
-      (event: any) => {
-        event.preventDefault();
-        spinner.down();
-        focusInput();
-        setIsAtMin(false);
-      },
-      [focusInput, spinner],
-    );
-
-    const spinStop = useCallback(
-      (event: any) => {
-        event.preventDefault();
-        spinner.stop();
-
-        if (isAtMinProp) {
-          setIsAtMin(true);
-        }
-      },
-      [isAtMinProp, spinner],
-    );
-
-    return {
-      tabIndex: -1,
-      onMouseDown: callAllHandlers(onMouseDown, spinDown),
-      onTouchStart: callAllHandlers(onTouchStart, spinDown),
-      onMouseLeave: callAllHandlers(onMouseUp, spinStop),
-      onMouseUp: callAllHandlers(onMouseUp, spinStop),
-      onTouchEnd: callAllHandlers(onTouchEnd, spinStop),
-      disabled: keepWithinRange && isAtMinProp && isAtMin,
-      "aria-disabled": ariaAttr(keepWithinRange && isAtMin),
-      ...restHtmlProps,
-    };
+    return useSpinButton(options, htmlProps, "decrement");
   },
 });
