+++ src/number-input/NumberInputIncrementButton.ts
@@ -1,8 +1,7 @@
-import { useCallback, useState } from "react";
 import { createComponent, createHook } from "reakit-system";
-import { ariaAttr, callAllHandlers } from "@chakra-ui/utils";
 import { ButtonHTMLProps, ButtonOptions, useButton } from "reakit/Button";
 
+import { useSpinButton } from "./__utils";
 import { NumberInputStateReturn } from "./NumberInputState";
 import { NUMBERINPUT_INCREMENTBUTTON_KEYS } from "./__keys";
 
@@ -27,56 +26,7 @@ export const useNumberInputIncrementButton = createHook<
   keys: NUMBERINPUT_INCREMENTBUTTON_KEYS,
 
   useProps(options, htmlProps) {
-    const {
-      keepWithinRange,
-      focusInput,
-      isAtMax: isAtMaxProp,
-      spinner,
-    } = options;
-    const {
-      onMouseDown,
-      onMouseUp,
-      onMouseLeave,
-      onTouchStart,
-      onTouchEnd,
-      ...restHtmlProps
-    } = htmlProps;
-
-    const [isAtMax, setIsAtMax] = useState(false);
-
-    const spinUp = useCallback(
-      (event: any) => {
-        event.preventDefault();
-        spinner.up();
-        focusInput();
-        setIsAtMax(false);
-      },
-      [focusInput, spinner],
-    );
-
-    const spinStop = useCallback(
-      (event: any) => {
-        event.preventDefault();
-        spinner.stop();
-
-        if (isAtMaxProp) {
-          setIsAtMax(true);
-        }
-      },
-      [isAtMaxProp, spinner],
-    );
-
-    return {
-      tabIndex: -1,
-      onMouseDown: callAllHandlers(onMouseDown, spinUp),
-      onTouchStart: callAllHandlers(onTouchStart, spinUp),
-      onMouseUp: callAllHandlers(onMouseUp, spinStop),
-      onMouseLeave: callAllHandlers(onMouseLeave, spinStop),
-      onTouchEnd: callAllHandlers(onTouchEnd, spinStop),
-      disabled: keepWithinRange && isAtMaxProp && isAtMax,
-      "aria-disabled": ariaAttr(keepWithinRange && isAtMax),
-      ...restHtmlProps,
-    };
+    return useSpinButton(options, htmlProps);
   },
 });
